requirements.txt
numpy
pandas
petname
psutil
torch
gpuinfo
matplotlib
twilio
PyYAML
scikit_learn
python.binance